Keep track of number of queries in the database ourselves. No need to fully count the number of queries in the disk after startup. Even though this is happening using a covering index, it turned out to be very slow on heavily I/O limited devices when the index tree does not fit into available cache memory (scanning can take upwards of one minute for a simple SELECT COUNT(*) FROM query_storage)

+++ src/api/info.c
@@ -148,13 +148,13 @@ int api_info_database(struct ftl_conn *api)
 
 	// Add number of queries and earliest timestamp in in-memory database
 	double earliest_timestamp_mem = 0.0;
-	const int queries_in_database = get_number_of_queries_in_DB(NULL, "query_storage", &earliest_timestamp_mem);
+	const int64_t queries_in_database = get_number_of_queries_in_DB(NULL, "query_storage", &earliest_timestamp_mem);
 	JSON_ADD_NUMBER_TO_OBJECT(json, "queries", queries_in_database);
 	JSON_ADD_NUMBER_TO_OBJECT(json, "earliest_timestamp", earliest_timestamp_mem);
 
 	// Add number of queries and earliest timestamp in on-disk database
 	double earliest_timestamp_disk = 0.0;
-	const int queries_in_disk_database = get_number_of_queries_in_DB(NULL, "disk.query_storage", &earliest_timestamp_disk);
+	const int64_t queries_in_disk_database = get_number_of_queries_in_DB(NULL, "disk.query_storage", &earliest_timestamp_disk);
 	JSON_ADD_NUMBER_TO_OBJECT(json, "queries_disk", queries_in_disk_database);
 	JSON_ADD_NUMBER_TO_OBJECT(json, "earliest_timestamp_disk", earliest_timestamp_disk);
